[feature] store refatoring page 완료
